Created rarities data table to simplify get upgrades function
I created a rarities data table which holds the different rarities in the game and the colour assigned to each rarity. This meant that I could simplify my code in get upgrade functions. This also allows me to easy add and remove rarities as its all now stored in a data table